requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
openai==1.41.1
mysql-connector-python==8.0.33
requests==2.28.1
python-dotenv==0.21.0
